a lot of shit
updated movement/camera code.

added the ability to automatically switch to the correct timeMechanic right click depending on which scene is currently loaded (to allow for easy use of switching time mechanic's.

Added code to spawn a timeBubble to increase the speed of objects thats able too be increased when inside the bubble (currently the movement part isn't working properly, but everything else is)