Main Menu
I have set up a really basic but mostly functional main menu system, the visual quality of this is a work in progress and the setting in the options menu are simply placeholder for now.
Start loads our demo scene, back goes back to main menu and quit exits the game. There is also an about game section whiith a brief description.