Updates
made changes to the boss so each mechanic is its own modular script.

Will make things alot easier in the future as it will allow plug and play mechanics to each boss.

also added an incomplete script to allow for the boss to teleport to specific locations as another mechanic