updated bullet piller
bullet pillar no longer allows bullets to shoot through it, accomplished this by adding a rigidbody and freezing its locations, this is not ideal for everything as this is more computationally expensive, but there is a better way by using ray sensors when shooting the bullet. but for now this solves that problem.