Teleport mechanic changes
made additional changes to the teleport script by adding a new function that'll allow for the boss to teleport to the further teleport location from its current position & added in additional data on the script in the editor itself to allow for customisation without having to go into the script itself.

currently you can decide on which two teleport methods you prefer, random teleport or furthest away teleport and the health at which you want the boss to teleport at.