Changes to health & shotgun & time gun
Added a health bar that decreases based on the health of the enemy and boss.

fixed the enemy prefab as it was running off the base health class (for some reason?) not its own custom enemy health script

Fixed the colour issue with the time gun & shotgun power up.

while I was there, made a change to the shotgun so it only spreads horizontally not vertically aswell anymore.